+++ Commands/about.js
@@ -19,7 +19,7 @@ module.exports = command;
 
 function creditsBuilder(footerURL, locale) {
 	return new MessageEmbed().setColor(`6b81eb`)
-		.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/bcE3Syu `)
+		.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/FJ8JGq2`)
 		.setTitle(getString(locale, command.module, "creditsTitle"))
 		.setThumbnail('https://cdn.discordapp.com/attachments/545684759276421120/734097014974971924/theater-curtains.png')
 		.setURL(`https://github.com/Imaginary-Horizons-Productions/DirectoryBot `)