+++ Commands/help.js
@@ -39,7 +39,7 @@ command.execute = (receivedMessage, state, locale) => {
 		commandSets.forEach(commandSet => {
 			if (state.botManager || !commandSet.managerCommand) {
 				var embed = new MessageEmbed().setColor('6b81eb')
-					.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/bcE3Syu `)
+					.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/FJ8JGq2`)
 					.setTitle(getString(locale, commandSet.module, "title"))
 					.setThumbnail('https://cdn.discordapp.com/attachments/545684759276421120/765059662268727326/info.png')
 					.setDescription(getString(locale, commandSet.module, "description"))