+++ Commands/lookup.js
@@ -7,7 +7,7 @@ var command = new Command("lookup", false, false, false);
 
 // Generate embed on call to add up-to-date list of platforms
 command.help = (avatarURL, guildID, locale, guildName, module) => {
-	let embed = new MessageEmbed().setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/bcE3Syu `)
+	let embed = new MessageEmbed().setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/FJ8JGq2`)
 		.setTitle(getString(locale, "DirectoryBot", "directoryBotCommand") + getString(locale, module, "names").join(', '))
 		.setDescription(getString(locale, module, "description"))
 		.setFooter(getString(locale, "DirectoryBot", "footerText"), avatarURL);