+++ Commands/support.js
@@ -19,7 +19,7 @@ module.exports = command;
 
 function supportBuilder(footerURL, locale) {
 	return new MessageEmbed().setColor('6b81eb')
-		.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/bcE3Syu `)
+		.setAuthor("Imaginary Horizons Productions", `https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png `, `https://discord.gg/FJ8JGq2`)
 		.setTitle(getString(locale, command.module, "supportingDirectoryBot"))
 		.setThumbnail(`https://cdn.discordapp.com/attachments/545684759276421120/734202424960745545/love-mystery.png`)
 		.setDescription(getString(locale, command.module, "embedDescription"))